got rid of the weird transition when opening the app
