+++ app/src/main/AndroidManifest.xml
@@ -17,16 +17,15 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.RememberMe">
         <activity android:name=".WalkthroughActivity">
-            <intent-filter>
-            <action android:name="android.intent.action.MAIN" />
-
-            <category android:name="android.intent.category.LAUNCHER" />
-        </intent-filter>
         </activity>
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
 
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
 
         <activity android:name=".EditFramilyProfile"></activity>
